Update atlantis from upstream (#64)
* Support using repo configuration from a branch

Instead of using the atlantis.yaml file present in the pull request
branch, allow users to specify a branch that contains an approved
version of atlantis.yaml. This allows enabling things like approval
requirement overrides for repo config, while ensuring a user can't
simply change the configuration to drop an approval requirement in the
branch they're currently working on.

An example would be creating a server-side configuration like this:

```yaml
---
repos:
  - id: github.com/gocardless/app
    apply_requirements: [approved, mergeable]
    allowed_overrides: [apply_requirements, workflow]
    allow_custom_workflows: false
    config_source_branch: master
workflows:
  lab:
    plan:
      steps:
        - init
        - plan:
            extra_args: [-var-file, lab.tfvars]
  prd:
    plan:
      steps:
        - init
        - plan:
            extra_args: [-var-file, prd.tfvars]
```

Specifying a rigid workflows that become all that is available to run in
gocardless/app. Now the application (gocardless/app) would create an
atlantis.yaml at the root of the repo, like so:

```
---
version: 2
projects:
  - dir: terraform/google/projects/apps
    apply_requirements: []
    workspace: lab
    workflow: lab
  - dir: terraform/google/projects/apps
    apply_requirements: [approved]
    workspace: prd
    workflow: prd
```

Because config_source_branch is set to master, when someone creates a PR
against gocardless/apps, the atlantis.yaml that specifies whether a
project has an approved apply_requirement is from the master branch. As
is a common pattern, merging to master is protected by an authorised
code-review, allowing us to have two types of environment for this
project:

- lab, where you can plan and apply without approval
- prd, where you can only apply once the PR is approved

* Bump go to v1.14 in docker

* Fixup int to string

Resolves this issue after upgrade to go 1.14:
server/events/db/boltdb.go:150:87: conversion from int to string yields a string of one rune, not a string of digits (did you mean fmt.Sprint(x)?)

Update hook calls to methods whose signature has changed

Disabled Dockerfile.dev, currently broken on upstream

* Bump up available patch versions of terraform

This should mimic the changes on the GC master branch

---------

Co-authored-by: Lawrence Jones <lawrjone@gmail.com>
Co-authored-by: Dyson Simmons <dysonsimmons@gmail.com>