Removed default terraform configuration
Atlantis server requires a default tf version, however, this setting
here also has the side effect to have a default version in utopia
terraform. This is not safe, as you might updating the tf state version
by mistake.

Move this setting to the atlantis runtime config, so we check and
run only the relevant terraform version in case of manual intervention