INC-5131: APIController: /api/apply pretend PRs have been approved
When a request to apply run is given from the /api/, we don't have
enough context regarding if the PR has actually been approved or not.

To work around this, we fake it, and pretend the PR is okay.
--
Note, this is _not_ the same code path that is used when the `atlantis apply`
comment is posted on a Github PR.

This only applies to runs that are triggered, bypassing the normal GH comment
process. This code can be disabled by unsetting ATLANTIS_API_SECRET to
disable these APIs.